!Templates/Dynamic/Crypto and all python/Curveball/files/requirements.txt
ecdsa
pycryptodome